requirements.txt
asgiref==3.7.2
boto3==1.26.153
botocore==1.29.153
certifi==2023.5.7
charset-normalizer==3.1.0
dj-database-url==2.0.0
Django==4.2.2
django-environ==0.10.0
django-on-heroku==1.1.2
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
psycopg2-binary==2.9.6
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==1.26.16
whitenoise==6.5.0
